refactor: option 단계 구분
